Inline EthHandler abstract class
