+++ rskj-core/src/main/java/org/ethereum/net/server/ChannelManagerImpl.java
@@ -30,7 +30,6 @@ import org.ethereum.config.NodeFilter;
 import org.ethereum.core.Block;
 import org.ethereum.core.BlockIdentifier;
 import org.ethereum.core.Transaction;
-import org.ethereum.net.eth.message.EthMessage;
 import org.ethereum.net.message.ReasonCode;
 import org.ethereum.sync.SyncPool;
 import org.slf4j.Logger;
